🧬 Migrate the data models about settings in section *<API>.http.response* to module *api_config.apis.response*.
